start refactoring Requests
