--- src/components/Request/index.ts
@@ -1,1 +1,0 @@
-export * from "./Request";
