+++ src/components/Requests/Request.scss
@@ -1,6 +1,6 @@
-@import "constants/style";
+@import "src/constants/style";
 
-.join-request {
+.requests__wrapper {
   position: fixed;
   min-width: 300px;
   max-width: 480px;
@@ -10,90 +10,27 @@
   right: 10px;
   overflow: hidden;
   animation: fly-in-top 0.5s ease-out;
-  box-shadow: 0 16px 32px 0 rgba(0, 87, 255, 0.24);
   z-index: $request-z-index;
 }
 
 @media screen and (min-width: 420px) {
-  .join-request {
+  .requests__wrapper {
     left: calc(50% - 200px);
     width: 480px;
   }
 }
 @media #{$tablet} {
-  .join-request {
+  .requests__wrapper {
     right: 92px;
     left: auto;
     animation: fly-in-right 0.5s ease-out;
-  }
-}
-[theme="dark"] {
-  .join-request {
-    box-shadow: 0 6px 9px 0 rgba(0, 0, 0, 0.2);
-  }
-}
-
-.request__button {
-  font-size: $text-size--medium;
-  font-weight: 600;
-  letter-spacing: $letter-spacing--small;
-  line-height: 24px;
-  border: none;
-  outline: none;
-  margin: 0;
-  padding: 0;
-  background-color: transparent;
-  color: $color-backlog-blue;
-  cursor: pointer;
-  transition: all 0.08s ease-out;
-
-  &:hover {
-    filter: brightness(1.4);
-    transform: scale(1.06);
-  }
-}
-
-.request__button + .request__button {
-  margin-left: $margin--default;
-}
-
-[theme="dark"] {
-  .request__button {
-    color: $color-planning-pink;
-  }
-}
-
-.request__header {
-  height: 42px;
-  font-size: $text-size--medium;
-  letter-spacing: $letter-spacing--small;
-  display: flex;
-  align-items: center;
-  font-weight: 600;
-  position: relative;
-  padding-left: $padding--default;
-  color: $color-white;
-  background-color: $color-backlog-blue;
-  overflow-wrap: break-word;
-  white-space: nowrap;
-}
-
-[theme="dark"] {
-  .request__header {
-    background-color: $color-planning-pink;
-  }
-}
-
-@media screen and (min-width: 375px) {
-  .request__header {
-    font-size: $text-size--medium;
   }
 }
 
 .request__main {
   width: 100%;
   height: auto;
-  background-color: $color-white;
+  // TODO: remove bg color and border, use those for list items
   border-radius: 0;
 }
 
@@ -107,6 +44,7 @@
   display: flex;
   align-items: center;
   justify-content: space-between;
+  background-color: $color-white;
   margin: 0 $margin--default;
 }
 
@@ -207,54 +145,17 @@
 }
 
 .request__requests {
+  display: flex;
+  flex-direction: column;
+  // row-gap: 50px;
   list-style-type: none;
   margin: 0;
   padding: $padding--default 0;
-  max-height: 130px;
   overflow-y: auto;
 }
 [theme="dark"] {
   .request__requests {
     border-color: rgba(212, 217, 221, 0.2);
-  }
-}
-
-.join-request__request-display-name {
-  height: 18px;
-  color: $color-black;
-  font-size: $text-size--medium;
-  letter-spacing: $letter-spacing--small;
-  line-height: $line-height--medium;
-  width: 144px;
-  margin: 0 $margin--default;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  overflow: hidden;
-  display: inline-block;
-}
-@media screen and (max-width: 375px) {
-  .join-request__request-display-name {
-    width: 96px;
-  }
-}
-[theme="dark"] {
-  .join-request__request-display-name {
-    color: $color-white;
-  }
-}
-
-.request__footer {
-  display: flex;
-  justify-content: flex-end;
-  padding: 0 $padding--default;
-
-  background-color: $color-white;
-  height: 48px;
-}
-
-[theme="dark"] {
-  .request__footer {
-    background-color: $color-dark-mode-note;
   }
 }
