+++ src/components/Requests/index.ts
@@ -1,0 +1,2 @@
+export * from "./Requests";
+export * from "./Request";
